groups login and signup under (auth) to use layout to remove redundant code
