+++ client/app/(auth)/login/page.tsx
@@ -1,0 +1,6 @@
+import { FormWrapper } from "@/components/form-wrapper";
+
+const LoginPage = () => {
+  return <FormWrapper formType="login" />;
+};
+export default LoginPage;
