+++ client/app/(auth)/signup/page.tsx
@@ -1,0 +1,6 @@
+import { FormWrapper } from "@/components/form-wrapper";
+const SignupPage = () => {
+  return <FormWrapper formType="register" />;
+};
+
+export default SignupPage;
